refactor(ingest): move the write path out of the service
The counterpart to the read-path extraction, and the more consequential
half: a retrieval bug returns the wrong answer to one query, while a
write-path bug is permanent, because it decides what the store will contain
from then on. None of it could be exercised without a gRPC request around
it.

internal/ingest now owns what gets stored: extraction, exact and semantic
folding, contradiction detection and supersedes, embedding, session and
entity links, and tag and meaning auto-linking. Store and Extract are its
two entry points. The service keeps the wire format and the arguments that
belong to it -- notably session_id, where the two RPCs deliberately differ:
Store rejects a malformed id, Extract warns and carries on rather than
discarding a whole conversation over one field. That judgement now sits at
the boundary where the malformed value arrives.

memory.go is 1,579 lines to 738. Between this and the retrieval move, the
file that held the whole engine now holds protobuf translation and the
handlers that own it.

Behaviour is unchanged and the golden suite says so: recall@10 0.909, MRR
0.828, identical to three decimals before and after, plus the graph
integration suite under -race.

Two tests moved with the code and got better for it. TestExtractEmbedsConcurrently
and TestExtractLinksSemanticallyRelatedMemories drove the pipeline through
gRPC to assert things about embedding concurrency and graph density; they
now call the engine directly, which is what the seam is for.

Closes #65
